refactor: UserHome renamed UserProfile
