+++ client/components/index.js
@@ -4,7 +4,7 @@
  * having to figure out which file they belong to!
  */
 export {default as Navbar} from './Navbar'
-export {default as UserHome} from './UserHome'
+export {default as UserProfile} from './UserProfile'
 export {Login, Signup} from './AuthForm'
 export {default as Camera} from './Camera'
 export {default as StartWorkout} from './StartWorkout'
