+++ client/routes.js
@@ -2,7 +2,7 @@ import React, {Component} from 'react'
 import {connect} from 'react-redux'
 import {withRouter, Route, Switch} from 'react-router-dom'
 import PropTypes from 'prop-types'
-import {Login, Signup, UserHome, StartWorkout} from './components'
+import {Login, Signup, UserProfile, StartWorkout} from './components'
 import {me} from './store'
 
 /**
@@ -24,7 +24,7 @@ class Routes extends Component {
         {isLoggedIn && (
           <Switch>
             {/* Routes placed here are only available after logging in */}
-            <Route path="/profile" component={UserHome} />
+            <Route path="/profile" component={UserProfile} />
             <Route path="/start" component={StartWorkout} />
           </Switch>
         )}
